rename `s3` driver to `object`
